fix: changed some datatypes to BigDecimal to be compatible with oracle db

+++ src/main/java/ir/maktabsharif/home_service/model/comment/Comment.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 @Entity
 @Table(name = "comments")
@@ -20,7 +21,7 @@ public class Comment extends BaseEntity {
 
     private String context;
 
-    private Double expertScore;
+    private BigDecimal expertScore;
 
     @OneToOne
     private Order order;
